Step 6: Create Auth Listener Component
